Resolving remaining conflicts in execution test suite

+++ test/libsolidity/semanticTests/array/dynamic_out_of_bounds_array_access.sol
@@ -1,21 +1,21 @@
 contract c {
-    uint256[] data;
+    uint[] data;
 
-    function enlarge(uint256 amount) public returns (uint256) {
+    function enlarge(uint amount) public returns (uint) {
         while (data.length < amount) data.push();
         return data.length;
     }
 
-    function set(uint256 index, uint256 value) public returns (bool) {
+    function set(uint index, uint value) public returns (bool) {
         data[index] = value;
         return true;
     }
 
-    function get(uint256 index) public returns (uint256) {
+    function get(uint index) public returns (uint) {
         return data[index];
     }
 
-    function length() public returns (uint256) {
+    function length() public returns (uint) {
         return data.length;
     }
 }
@@ -24,11 +24,11 @@ contract c {
 // compileViaYul: also
 // ----
 // length() -> 0
-// get(uint256): 3 -> FAILURE, hex"4e487b71", 0x32
-// enlarge(uint256): 4 -> 4
+// get(uint): 3 -> FAILURE, hex"4e487b71", 0x32
+// enlarge(uint): 4 -> 4
 // length() -> 4
-// set(uint256,uint256): 3, 4 -> true
-// get(uint256): 3 -> 4
+// set(uint,uint): 3, 4 -> true
+// get(uint): 3 -> 4
 // length() -> 4
-// set(uint256,uint256): 4, 8 -> FAILURE, hex"4e487b71", 0x32
+// set(uint,uint): 4, 8 -> FAILURE, hex"4e487b71", 0x32
 // length() -> 4
