+++ test/libsolidity/semanticTests/array/fixed_arrays_in_storage.sol
@@ -1,34 +1,34 @@
 contract c {
     struct Data {
-        uint256 x;
-        uint256 y;
+        uint x;
+        uint y;
     }
     Data[2**10] data;
-    uint256[2**10 + 3] ids;
+    uint[2**10 + 3] ids;
 
-    function setIDStatic(uint256 id) public {
+    function setIDStatic(uint id) public {
         ids[2] = id;
     }
 
-    function setID(uint256 index, uint256 id) public {
+    function setID(uint index, uint id) public {
         ids[index] = id;
     }
 
-    function setData(uint256 index, uint256 x, uint256 y) public {
+    function setData(uint index, uint x, uint y) public {
         data[index].x = x;
         data[index].y = y;
     }
 
-    function getID(uint256 index) public returns (uint256) {
+    function getID(uint index) public returns (uint) {
         return ids[index];
     }
 
-    function getData(uint256 index) public returns (uint256 x, uint256 y) {
+    function getData(uint index) public returns (uint x, uint y) {
         x = data[index].x;
         y = data[index].y;
     }
 
-    function getLengths() public returns (uint256 l1, uint256 l2) {
+    function getLengths() public returns (uint l1, uint l2) {
         l1 = data.length;
         l2 = ids.length;
     }
@@ -37,12 +37,12 @@ contract c {
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// setIDStatic(uint256): 0xb ->
-// getID(uint256): 0x2 -> 0xb
-// setID(uint256,uint256): 0x7, 0x8 ->
-// getID(uint256): 0x7 -> 0x8
-// setData(uint256,uint256,uint256): 0x7, 0x8, 0x9 ->
-// setData(uint256,uint256,uint256): 0x8, 0xa, 0xb ->
-// getData(uint256): 0x7 -> 0x8, 0x9
-// getData(uint256): 0x8 -> 0xa, 0xb
+// setIDStatic(uint): 0xb ->
+// getID(uint): 0x2 -> 0xb
+// setID(uint,uint): 0x7, 0x8 ->
+// getID(uint): 0x7 -> 0x8
+// setData(uint,uint,uint): 0x7, 0x8, 0x9 ->
+// setData(uint,uint,uint): 0x8, 0xa, 0xb ->
+// getData(uint): 0x7 -> 0x8, 0x9
+// getData(uint): 0x8 -> 0xa, 0xb
 // getLengths() -> 0x400, 0x403
