+++ test/libsolidity/semanticTests/enums/enum_explicit_overflow.sol
@@ -3,19 +3,19 @@ contract test {
 
     constructor() {}
 
-    function getChoiceExp(uint256 x) public returns (uint256 d) {
+    function getChoiceExp(uint x) public returns (uint d) {
         choice = ActionChoices(x);
-        d = uint256(choice);
+        d = uint(choice);
     }
 
-    function getChoiceFromSigned(int256 x) public returns (uint256 d) {
+    function getChoiceFromSigned(int x) public returns (uint d) {
         choice = ActionChoices(x);
-        d = uint256(choice);
+        d = uint(choice);
     }
 
-    function getChoiceFromMax() public returns (uint256 d) {
+    function getChoiceFromMax() public returns (uint d) {
         choice = ActionChoices(type(uint).max);
-        d = uint256(choice);
+        d = uint(choice);
     }
 
     ActionChoices choice;
@@ -25,9 +25,9 @@ contract test {
 // compileViaYul: also
 // EVMVersion: >=byzantium
 // ----
-// getChoiceExp(uint256): 2 -> 2
-// getChoiceExp(uint256): 3 -> FAILURE, hex"4e487b71", 33 # These should throw #
-// getChoiceFromSigned(int256): -1 -> FAILURE, hex"4e487b71", 33
+// getChoiceExp(uint): 2 -> 2
+// getChoiceExp(uint): 3 -> FAILURE, hex"4e487b71", 33 # These should throw #
+// getChoiceFromSigned(int): -1 -> FAILURE, hex"4e487b71", 33
 // getChoiceFromMax() -> FAILURE, hex"4e487b71", 33
-// getChoiceExp(uint256): 2 -> 2 # These should work #
-// getChoiceExp(uint256): 0 -> 0
+// getChoiceExp(uint): 2 -> 2 # These should work #
+// getChoiceExp(uint): 0 -> 0
