+++ test/libsolidity/semanticTests/modifiers/function_modifier_multiple_times_local_vars.sol
@@ -1,18 +1,18 @@
 contract C {
-    uint256 public a;
-    modifier mod(uint256 x) {
-        uint256 b = x;
+    uint public a;
+    modifier mod(uint x) {
+        uint b = x;
         a += b;
         _;
         a -= b;
         assert(b == x);
     }
 
-    function f(uint256 x) public mod(2) mod(5) mod(x) returns (uint256) {
+    function f(uint x) public mod(2) mod(5) mod(x) returns (uint) {
         return a;
     }
 }
 
 // ----
-// f(uint256): 3 -> 10
+// f(uint): 3 -> 10
 // a() -> 0
